ci(example): migrate Patrol e2e/conformance to patrol_plus + cross-platform coverage
Switch the example app and CI off the old Patrol git fork
(github.com/ahmednfwela/patrol#fixed) onto the published patrol_plus /
patrol_cli_plus packages, and run the conformance + e2e + offline suites
under `patrol_plus test --coverage` on android/iOS/web/linux/windows.
macOS stays on `flutter test integration_test --coverage` (Patrol coverage
is unsupported on macOS).

example/pubspec.yaml:
- dev_dep: replace the `patrol` git fork with `patrol_plus: ^5.0.1` (pub).
- drop flutter_driver (web no longer uses flutter-drive/DWDS).
- keep integration_test + flutter_test for the macOS job.

tests:
- patrol_test/app_test.dart: import package:patrol_plus/patrol.dart.
- add integration_test/shared_offline.dart: harness-agnostic offline-mode
  logic parameterized by launch+pump callbacks.
- integration_test/offline_mode_test.dart now delegates to it (macOS path).
- add patrol_test/offline_mode_test.dart so the offline suite also runs
  under Patrol on android/iOS/linux/windows.

.github/workflows/tests.yaml:
- android/iOS/linux/windows: `dart pub global activate patrol_cli_plus` +
  `patrol_plus test --coverage --coverage-ignore="**/*.g.dart"
  --coverage-package oidc`, copying coverage/patrol_lcov.info to the
  per-platform <platform>-coverage.info the combine step expects.
- web: activate patrol_cli_plus (not the git fork), keep Playwright runner,
  best-effort V8 coverage upload.
- combine step now also waits on web.

linux/windows generated_plugins.cmake: register the patrol_plus desktop
native plugin (auto-generated by flutter pub get).
