+++ packages/oidc/example/windows/flutter/generated_plugins.cmake
@@ -5,6 +5,7 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   flutter_secure_storage_windows
   oidc_windows
+  patrol_plus
   url_launcher_windows
   window_to_front
 )
